requirements.txt
insightface==0.7.3
numpy==1.24.3
opencv-python-headless==4.8.1.78
Pillow==10.0.0
